requirements.txt
flask
gunicorn
pytest
psycopg2
pylint
pytest-flask
Flask-JSONRPC
Authlib
blinker
flask_cors
python-memcached
flask_sqlalchemy
flask_script
flask_migrate
flask-wtf
WTForms-Alchemy
